+++ src/main/kotlin/com/github/insanusmokrassar/SauceNaoAPI/models/LimitsState.kt
@@ -1,0 +1,11 @@
+package com.github.insanusmokrassar.SauceNaoAPI.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LimitsState(
+    val maxShortQuota: Int,
+    val maxLongQuota: Int,
+    val knownShortQuota: Int,
+    val knownLongQuota: Int
+)