+++ src/main/kotlin/com/github/insanusmokrassar/SauceNaoAPI/utils/TimeManager.kt
@@ -68,7 +68,7 @@ private data class TimeManagerMostOldestInShortGetter(
     }
 }
 
-class TimeManager(
+internal class TimeManager(
     scope: CoroutineScope
 ) : Closeable {
     private val actionsChannel = Channel<TimeManagerAction>(Channel.UNLIMITED)